refactor: added ThemeProvider & NotesProvier
shifted the state from app.jsx to global using context API so that it can be using in various components of the app without having to prop drilling
