+++ src/contexts/index.js
@@ -1,0 +1,2 @@
+export {NotesProvider} from "./notes-context";
+export {ThemeProvider} from "./theme-context";
