+++ src/index.js
@@ -3,9 +3,15 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 
+import {NotesProvider, ThemeProvider} from "./contexts";
+
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <ThemeProvider>
+      <NotesProvider>
+        <App />
+      </NotesProvider>
+    </ThemeProvider>
   </React.StrictMode>,
   document.getElementById('root')
 );
